// js/inc/back/attempt01/hw/server04/src/models/common.ts
import { Request } from "express";

export enum Status {
  Ok_200 = 200,
  Created_201 = 201,
  NoContent_204 = 204,
  BadRequest_400 = 400,
  Unauthorized_401 = 401,
  NotFound_404 = 404,
}

export interface IError {
  errorsMessages: IErrorMessage[];
}
export interface IErrorMessage {
  message: string;
  field: string;
}

// Requset
export type RequestWithBody<B> = Request<{}, {}, B>;
export type RequestWithParams<P> = Request<P>;
export type RequestWithBodyAndParams<P, B> = Request<P, {}, B>;
export type RequestWithQuery<Q> = Request<{}, {}, {}, Q>;
export type RequestWithParamsAndQuery<P, Q> = Request<P, {}, {}, Q>;

export type Pagination<T> = {
  pagesCount: number;
  page: number;
  pageSize: number;
  totalCount: number;
  items: T[];
};